requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
pytesseract==0.3.10
pdf2image==1.16.3
pandas==2.1.1
numpy==1.24.3
Pillow==10.0.1